modified message struct

+++ models/connection/message.go
@@ -2,12 +2,12 @@ package connection
 
 type NoPayload bool
 type Message[T any] struct {
-	Code    int      `json:"code"`
+	Code    uint8      `json:"code"`
 	Payload T        `json:"payload,omitempty"`
 	Error   *RespErr `json:"error,omitempty"`
 }
 
-func NewMessage[T any](code int) Message[T] {
+func NewMessage[T any](code uint8) Message[T] {
 	return Message[T]{Code: code}
 }
